Update devise, devise-encryptable
Upgrades the locked versions of the `devise` and `devise-encryptable` gems.

Devise has been bumped to a new major version, which removes support for
the following:

- Rails 3.2. and 4.0
- Ruby 1.9 and 2.0

These losses are deemed acceptable, as Solidus' core itself does not
support any of these versions.

For more details on the changes, see the Devise changelog:
https://github.com/plataformatec/devise/blob/master/CHANGELOG.md

Similar changes in `devise-encryptable`, with details in the changelog:
https://github.com/plataformatec/devise-encryptable/blob/master/Changelog.md